moved pag to main-page-only folder
