+++ src/index.js
@@ -4,4 +4,4 @@ import './js/render-markup';
 import './js/modal';
 
 import './js/search';
-import './js/pagination';
+import './js/main-page-only/pagination';
